requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
decorator==5.1.1
discord.py==2.3.2
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
idna==3.6
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
multidict==6.0.4
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.0
six==1.16.0
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.0
wcwidth==0.2.12
yarl==1.9.4
